Revert detect-builders and detect-routes changes
