+++ packages/now-build-utils/src/detect-builders.ts
@@ -1,12 +1,6 @@
 import minimatch from 'minimatch';
 import { valid as validSemver } from 'semver';
-import {
-  Builder,
-  Config,
-  BuilderFunctions,
-  DetectorResult,
-  DetectorOutput,
-} from './types';
+import { PackageJson, Builder, Config, BuilderFunctions } from './types';
 
 interface ErrorResponse {
   code: string;
@@ -16,6 +10,26 @@ interface ErrorResponse {
 interface Options {
   tag?: 'canary' | 'latest' | string;
   functions?: BuilderFunctions;
+}
+
+const src = 'package.json';
+const config: Config = { zeroConfig: true };
+
+const MISSING_BUILD_SCRIPT_ERROR: ErrorResponse = {
+  code: 'missing_build_script',
+  message:
+    'Your `package.json` file is missing a `build` property inside the `scripts` property.' +
+    '\nMore details: https://zeit.co/docs/v2/platform/frequently-asked-questions#missing-build-script',
+};
+
+// Static builders are special cased in `@now/static-build`
+function getBuilders({ tag }: Options = {}): Map<string, Builder> {
+  const withTag = tag ? `@${tag}` : '';
+  const config = { zeroConfig: true };
+
+  return new Map<string, Builder>([
+    ['next', { src, use: `@now/next${withTag}`, config }],
+  ]);
 }
 
 // Must be a function to ensure that the returned
@@ -33,8 +47,13 @@ function getApiBuilders({ tag }: Pick<Options, 'tag'> = {}): Builder[] {
   ];
 }
 
-function hasDirectory(fileName: string, files: string[]) {
-  return files.some(name => name.startsWith(`${fileName}/`));
+function hasPublicDirectory(files: string[]) {
+  return files.some(name => name.startsWith('public/'));
+}
+
+function hasBuildScript(pkg: PackageJson | undefined) {
+  const { scripts = {} } = pkg || {};
+  return Boolean(scripts && scripts['build']);
 }
 
 function getApiFunctionBuilder(
@@ -72,71 +91,39 @@ function getApiFunctionBuilder(
   return use ? { use, src, config } : prevBuilder;
 }
 
-function detectFrontBuilder(
-  detectorResult: Partial<DetectorOutput>,
+async function detectFrontBuilder(
+  pkg: PackageJson,
   builders: Builder[],
-  files: string[],
   options: Options
-): Builder {
+): Promise<Builder> {
   const { tag } = options;
   const withTag = tag ? `@${tag}` : '';
+  for (const [dependency, builder] of getBuilders(options)) {
+    const deps = Object.assign({}, pkg.dependencies, pkg.devDependencies);
 
-  const {
-    framework,
-    buildCommand,
-    outputDirectory,
-    devCommand,
-  } = detectorResult;
-
-  const frameworkSlug = framework ? framework.slug : null;
-
-  const config: Config = {
-    zeroConfig: true,
-  };
-
-  if (framework) {
-    config.framework = framework;
-  }
-
-  if (devCommand) {
-    config.devCommand = devCommand;
-  }
-
-  if (buildCommand) {
-    config.buildCommand = buildCommand;
-  }
-
-  if (outputDirectory) {
-    config.outputDirectory = outputDirectory;
-  }
-
-  // All unused functions will be used for the frontend
-  if (options.functions) {
-    Object.entries(options.functions).forEach(([key, func]) => {
-      if (
-        builders.every(
-          b => !(b.config && b.config.functions && b.config.functions[key])
-        )
-      ) {
-        config.functions = config.functions || {};
-        config.functions[key] = { ...func };
+    // Return the builder when a dependency matches
+    if (deps[dependency]) {
+      if (options.functions) {
+        Object.entries(options.functions).forEach(([key, func]) => {
+          // When the builder is not used yet we'll use it for the frontend
+          if (
+            builders.every(
+              b => !(b.config && b.config.functions && b.config.functions[key])
+            )
+          ) {
+            if (!builder.config) builder.config = {};
+            if (!builder.config.functions) builder.config.functions = {};
+            builder.config.functions[key] = { ...func };
+          }
+        });
       }
-    });
-  }
 
-  if (frameworkSlug === 'next') {
-    return { src: 'package.json', use: `@now/next${withTag}`, config };
-  }
-
-  if (frameworkSlug === 'hugo') {
-    const configFiles = new Set(['config.yaml', 'config.toml', 'config.json']);
-    const source = files.find(file => configFiles.has(file));
-    if (source) {
-      return { src: source, use: `@now/static-build${withTag}`, config };
+      return builder;
     }
   }
 
-  return { src: 'package.json', use: `@now/static-build${withTag}`, config };
+  // By default we'll choose the `static-build` builder
+  return { src, use: `@now/static-build${withTag}`, config };
 }
 
 // Files that match a specific pattern will get ignored
@@ -152,6 +139,10 @@ export function getIgnoreApiFilter(optionsOrBuilders: Options | Builder[]) {
   }
 
   return (file: string) => {
+    if (!file.startsWith('api/')) {
+      return false;
+    }
+
     if (file.includes('/.')) {
       return false;
     }
@@ -178,7 +169,10 @@ export function sortFiles(fileA: string, fileB: string) {
   return fileA.localeCompare(fileB);
 }
 
-function detectApiBuilders(files: string[], options: Options): Builder[] {
+async function detectApiBuilders(
+  files: string[],
+  options: Options
+): Promise<Builder[]> {
   const builds = files
     .sort(sortFiles)
     .filter(getIgnoreApiFilter(options))
@@ -194,10 +188,10 @@ function detectApiBuilders(files: string[], options: Options): Builder[] {
 
 // When a package has files that conflict with `/api` routes
 // e.g. Next.js pages/api we'll check it here and return an error.
-function checkConflictingFiles(
+async function checkConflictingFiles(
   files: string[],
   builders: Builder[]
-): ErrorResponse | null {
+): Promise<ErrorResponse | null> {
   // For Next.js
   if (builders.some(b => b.use.startsWith('@now/next'))) {
     const hasApiPages = files.some(file => file.startsWith('pages/api/'));
@@ -217,10 +211,10 @@ function checkConflictingFiles(
 
 // When e.g. Next.js receives a `functions` property it has to make sure,
 // that it can handle those files, otherwise there are unused functions.
-function checkUnusedFunctionsOnFrontendBuilder(
+async function checkUnusedFunctionsOnFrontendBuilder(
   files: string[],
   builder: Builder
-): ErrorResponse | null {
+): Promise<ErrorResponse | null> {
   const { config: { functions = undefined } = {} } = builder;
 
   if (!functions) return null;
@@ -237,9 +231,7 @@ function checkUnusedFunctionsOnFrontendBuilder(
       ) {
         return {
           code: 'unused_function',
-          message:
-            `The function for "${matchedFile}" can't be handled by any runtime. ` +
-            `Please provide one with the "runtime" option.`,
+          message: `The function for ${matchedFile} can't be handled by any builder`,
         };
       }
     }
@@ -249,6 +241,8 @@ function checkUnusedFunctionsOnFrontendBuilder(
 }
 
 function validateFunctions(files: string[], { functions = {} }: Options) {
+  const apiBuilders = getApiBuilders();
+
   for (const [path, func] of Object.entries(functions)) {
     if (path.length > 256) {
       return {
@@ -318,6 +312,15 @@ function validateFunctions(files: string[], { functions = {} }: Options) {
             'Function Runtimes must have a valid version, for example `now-php@1.0.0`.',
         };
       }
+
+      if (
+        apiBuilders.some(b => func.runtime && func.runtime.startsWith(b.use))
+      ) {
+        return {
+          code: 'invalid_function_runtime',
+          message: `The function Runtime ${func.runtime} is not a Community Runtime and must not be specified.`,
+        };
+      }
     }
 
     if (func.includeFiles !== undefined) {
@@ -346,7 +349,7 @@ function validateFunctions(files: string[], { functions = {} }: Options) {
 // to determine what builders to use
 export async function detectBuilders(
   files: string[],
-  detectorResult: Partial<DetectorResult> | null = null,
+  pkg?: PackageJson | undefined | null,
   options: Options = {}
 ): Promise<{
   builders: Builder[] | null;
@@ -367,24 +370,19 @@ export async function detectBuilders(
   }
 
   // Detect all builders for the `api` directory before anything else
-  const builders = detectApiBuilders(files, options);
+  const builders = await detectApiBuilders(files, options);
 
-  if (detectorResult && detectorResult.buildCommand) {
-    const frontendBuilder = detectFrontBuilder(
-      detectorResult,
-      builders,
-      files,
-      options
-    );
+  if (pkg && hasBuildScript(pkg)) {
+    const frontendBuilder = await detectFrontBuilder(pkg, builders, options);
     builders.push(frontendBuilder);
 
-    const conflictError = checkConflictingFiles(files, builders);
+    const conflictError = await checkConflictingFiles(files, builders);
 
     if (conflictError) {
       warnings.push(conflictError);
     }
 
-    const unusedFunctionError = checkUnusedFunctionsOnFrontendBuilder(
+    const unusedFunctionError = await checkUnusedFunctionsOnFrontendBuilder(
       files,
       frontendBuilder
     );
@@ -396,35 +394,34 @@ export async function detectBuilders(
         warnings,
       };
     }
-  } else if (
-    detectorResult &&
-    detectorResult.outputDirectory &&
-    hasDirectory(detectorResult.outputDirectory, files)
-  ) {
-    builders.push({
-      use: '@now/static',
-      src: [...detectorResult.outputDirectory.split('/'), '**', '*']
-        .filter(Boolean)
-        .join('/'),
-      config: { zeroConfig: true },
-    });
-  } else if (hasDirectory('public', files)) {
-    builders.push({
-      use: '@now/static',
-      src: 'public/**/*',
-      config: { zeroConfig: true },
-    });
-  } else if (
-    builders.length > 0 &&
-    files.some(f => !f.startsWith('api/') && f !== 'package.json')
-  ) {
-    // Everything besides the api directory
-    // and package.json can be served as static files
-    builders.push({
-      use: '@now/static',
-      src: '!{api/**,package.json}',
-      config: { zeroConfig: true },
-    });
+  } else {
+    if (pkg && builders.length === 0) {
+      // We only show this error when there are no api builders
+      // since the dependencies of the pkg could be used for those
+      errors.push(MISSING_BUILD_SCRIPT_ERROR);
+      return { errors, warnings, builders: null };
+    }
+
+    // We allow a `public` directory
+    // when there are no build steps
+    if (hasPublicDirectory(files)) {
+      builders.push({
+        use: '@now/static',
+        src: 'public/**/*',
+        config,
+      });
+    } else if (
+      builders.length > 0 &&
+      files.some(f => !f.startsWith('api/') && f !== 'package.json')
+    ) {
+      // Everything besides the api directory
+      // and package.json can be served as static files
+      builders.push({
+        use: '@now/static',
+        src: '!{api/**,package.json}',
+        config,
+      });
+    }
   }
 
   return {
